test/naiverag_code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.16=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
asgiref=3.8.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=25.3.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.3.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py310h6a678d5_9
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2025.1.31=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=1.0.4=pypi_0
click=8.1.8=pypi_0
coloredlogs=15.0.1=pypi_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.4.127=0
cuda-runtime=12.4.1=0
cuda-version=11.8=h70ddcb2_3
cudatoolkit=11.8.0=h4ba93d1_13
dashscope=1.23.3=pypi_0
dataclasses-json=0.6.7=pypi_0
deprecated=1.2.18=pypi_0
distro=1.9.0=pypi_0
durationpy=0.9=pypi_0
exceptiongroup=1.2.2=pypi_0
faiss=1.9.0=py310cuda118hce52afa_0_cuda
faiss-gpu=1.9.0=h418687c_0
fastapi=0.115.9=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.17.0=py310h06a4308_0
flatbuffers=25.2.10=pypi_0
freetype=2.13.3=h4a9f257_0
frozenlist=1.5.0=pypi_0
fsspec=2025.3.2=pypi_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py310h5eee18b_0
gnutls=3.6.15=he1e5248_0
google-auth=2.38.0=pypi_0
googleapis-common-protos=1.69.2=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.67.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.27.2=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.30.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=py310h06a4308_0
importlib-metadata=8.6.1=pypi_0
importlib-resources=6.5.2=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=py310h06a4308_0
jiter=0.9.0=pypi_0
joblib=1.5.1=pypi_0
jpeg=9e=h5eee18b_3
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
kernel-headers_linux-64=3.10.0=he073ed8_18
kubernetes=32.0.1=pypi_0
lame=3.100=h7b6447c_0
langchain=0.3.23=pypi_0
langchain-community=0.3.21=pypi_0
langchain-core=0.3.65=pypi_0
langchain-huggingface=0.3.0=pypi_0
langchain-openai=0.3.12=pypi_0
langchain-text-splitters=0.3.8=pypi_0
langchainhub=0.1.21=pypi_0
langsmith=0.3.45=pypi_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libblas=3.9.0=1_h86c2bf4_netlib
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libdeflate=1.22=h5eee18b_0
libfaiss=1.9.0=cuda118hf323835_0_cuda
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgfortran=14.2.0=h69a702a_2
libgfortran-ng=14.2.0=h69a702a_2
libgfortran5=14.2.0=hf1ad2bd_2
libgomp=14.2.0=h767d61c_2
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=8_h3b12eaf_netlib
libnpp=12.2.5.30=0
libnvfatbin=12.4.127=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libpng=1.6.39=h5eee18b_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=14.2.0=h4852527_2
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py310h5eee18b_0
marshmallow=3.26.1=pypi_0
mdurl=0.1.2=pypi_0
milvus-lite=2.4.12=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mmh3=5.1.0=pypi_0
monotonic=1.6=pypi_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py310h06a4308_0
multidict=6.4.3=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=py310h06a4308_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
oauthlib=3.2.2=pypi_0
ocl-icd=2.3.2=h5eee18b_1
ollama=0.4.8=pypi_0
onnxruntime=1.21.0=pypi_0
openai=1.72.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.5.0=h7b32b05_0
opentelemetry-api=1.32.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.32.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.32.0=pypi_0
opentelemetry-instrumentation=0.53b0=pypi_0
opentelemetry-instrumentation-asgi=0.53b0=pypi_0
opentelemetry-instrumentation-fastapi=0.53b0=pypi_0
opentelemetry-proto=1.32.0=pypi_0
opentelemetry-sdk=1.32.0=pypi_0
opentelemetry-semantic-conventions=0.53b0=pypi_0
opentelemetry-util-http=0.53b0=pypi_0
orjson=3.10.16=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
pillow=11.1.0=py310hcea889d_0
pip=25.0=py310h06a4308_0
posthog=3.24.0=pypi_0
propcache=0.3.1=pypi_0
protobuf=5.29.4=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
pydantic-settings=2.8.1=pypi_0
pygments=2.19.1=pypi_0
pymilvus=2.5.6=pypi_0
pymilvus-model=0.3.2=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.0=pypi_0
python_abi=3.10=2_cp310
pytorch=2.5.1=py3.10_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2025.2=pypi_0
pyyaml=6.0.2=py310h5eee18b_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=py310h06a4308_1
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=14.0.0=pypi_0
rpds-py=0.24.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.7.0=pypi_0
scipy=1.15.2=pypi_0
sentence-transformers=4.1.0=pypi_0
setuptools=75.8.0=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.45.3=pypi_0
sympy=1.13.1=pypi_0
sysroot_linux-64=2.17=h0157908_18
tbb=2021.8.0=hdb19cb5_0
tenacity=9.1.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.9.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.1=pypi_0
tomli=2.2.1=pypi_0
torchaudio=2.5.1=py310_cu124
torchtriton=3.1.0=py310
torchvision=0.20.1=py310_cu124
tqdm=4.67.1=pypi_0
transformers=4.51.3=pypi_0
typer=0.15.2=pypi_0
types-requests=2.32.0.20250328=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.0=pypi_0
typing_extensions=4.12.2=py310h06a4308_0
tzdata=2025.2=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.3.0=py310h06a4308_0
uvicorn=0.34.0=pypi_0
uvloop=0.21.0=pypi_0
watchfiles=1.0.5=pypi_0
websocket-client=1.8.0=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py310h06a4308_0
wrapt=1.17.2=pypi_0
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.19.0=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=pypi_0
zstd=1.5.6=hc292b87_0
